backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.1.0
python-multipart==0.0.6
feedparser>=6.0.11
newspaper3k==0.2.8
lxml_html_clean>=0.1.0
rapidfuzz==3.6.1
cachetools==5.3.2
langchain>=0.3.0
langchain-groq>=0.1.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
sentence-transformers==2.3.1
faiss-cpu>=1.9.0
edge-tts==6.1.9
httpx==0.26.0
python-dotenv==1.0.0
websockets==12.0
